# business/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8082
info:
  contact:
    email: dev@watink.com.br
    name: Watink Dev
  description: API do backend Go — atendimento e automação WhatsApp.
  license:
    name: Proprietary
    url: https://watink.com.br
  termsOfService: https://watink.com.br/terms
  title: Watink Business API
  version: 2.0.0
paths:
...
  /pipelines:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar pipelines
      tags:
      - pipelines
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do pipeline
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar pipeline
      tags:
      - pipelines
  /pipelines/{pipelineId}:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do pipeline
        in: path
        name: pipelineId
        required: true
        type: integer
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar pipeline
      tags:
      - pipelines
  /pipelines/ai-suggest:
    post:
      consumes:
      - application/json
      description: Gera sugestão de pipeline com IA baseada em contexto de atendimento
      parameters:
      - description: Mensagens de contexto
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Sugestão IA de pipeline
      tags:
      - pipelines
  /pipelines/export/{pipelineId}:
    get:
      parameters:
      - description: ID do pipeline
        in: path
        name: pipelineId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Exportar pipeline
      tags:
      - pipelines
  /pipelines/import:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do pipeline
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Importar pipeline
      tags:
      - pipelines
...
  /queue:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar filas
      tags:
      - queues
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da fila
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar fila
      tags:
      - queues
  /queue/{queueId}:
    delete:
      parameters:
      - description: ID da fila
        in: path
        name: queueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remover fila
      tags:
      - queues
    get:
      parameters:
      - description: ID da fila
        in: path
        name: queueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Detalhar fila
      tags:
      - queues
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da fila
        in: path
        name: queueId
        required: true
        type: integer
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar fila
      tags:
      - queues
...
  /system/rabbitmq/queues:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Filas RabbitMQ
      tags:
      - system
  /system/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Estatísticas do sistema
      tags:
      - system
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token. Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"